test(ai): AI Gateway client 골격 테스트 추가
